+++ CHANGELOG.rst
@@ -3,7 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
 - (Fixed) YAML values like "1", "1.0", "yes", "True" are converted to strings, not to in, float and bool respectively.
+- (Changed) ``ruamel.yaml`` replaces ``pyyaml`` as supported YAML parser.
 
 
 0.22.7 (2016-08-19)