+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-Unreleased
-----------
+0.22.8 (2016-10-19)
+-------------------
 
 - (Fixed) YAML values like "1", "1.0", "yes", "True" are converted to strings, not to in, float and bool respectively.
 - (Changed) ``ruamel.yaml`` replaces ``pyyaml`` as supported YAML parser.